+++ Test/Tasty/Lens/Prism.hs
@@ -1,6 +1,6 @@
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE RankNTypes #-}
-module Test.Tasty.SmallCheck.Lens.Prism where
+module Test.Tasty.Lens.Prism where
 
 import Control.Lens
 import Test.SmallCheck.Series (Serial(series), CoSerial)