+++ Test/Tasty/Lens/Traversal.hs
@@ -1,7 +1,7 @@
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE RankNTypes #-}
 {-# LANGUAGE ScopedTypeVariables #-}
-module Test.Tasty.SmallCheck.Lens.Traversal where
+module Test.Tasty.Lens.Traversal where
 
 import Control.Lens
 import Test.SmallCheck.Series (Serial(series), CoSerial, Series)